fixed circular dependency + notes
