+++ src/stories/0_Notes.stories.mdx
@@ -15,6 +15,9 @@ The following components are pending deprecation and will be removed in next rel
 
 <br /><br />
 
+## 1.12.1
+- Fixed circular dependency in Popover component
+
 ## 1.12.0
 - Added [useConfirmationModal](?path=/docs/utils-confirmation-modal--page) utility hook
